+++ crates/store/src/errors.rs
@@ -1,16 +1,11 @@
 use std::io;
 
 use deadpool::managed::PoolError;
-use deadpool_sync::InteractError;
 use miden_objects::{
-    AccountDeltaError, AccountError, NoteError,
+    AccountDeltaError, AccountError, AccountTreeError, NoteError, NullifierTreeError,
     account::AccountId,
-    block::{BlockHeader, BlockNumber},
-    crypto::{
-        hash::rpo::RpoDigest,
-        merkle::{MerkleError, MmrError},
-        utils::DeserializationError,
-    },
+    block::BlockNumber,
+    crypto::{hash::rpo::RpoDigest, merkle::MmrError, utils::DeserializationError},
     note::Nullifier,
     transaction::OutputNote,
 };
@@ -18,18 +13,6 @@ use rusqlite::types::FromSqlError;
 use thiserror::Error;
 use tokio::sync::oneshot::error::RecvError;
 use tonic::Status;
-
-// INTERNAL ERRORS
-// =================================================================================================
-
-#[derive(Debug, Error)]
-pub enum NullifierTreeError {
-    #[error("failed to create nullifier tree")]
-    CreationFailed(#[source] MerkleError),
-
-    #[error("failed to mutate nullifier tree")]
-    MutationFailed(#[source] MerkleError),
-}
 
 // DATABASE ERRORS
 // =================================================================================================
@@ -109,21 +92,21 @@ pub enum StateInitializationError {
     #[error("failed to create nullifier tree")]
     FailedToCreateNullifierTree(#[from] NullifierTreeError),
     #[error("failed to create accounts tree")]
-    FailedToCreateAccountsTree(#[from] MerkleError),
+    FailedToCreateAccountsTree(#[source] AccountTreeError),
 }
 
 #[derive(Debug, Error)]
 pub enum DatabaseSetupError {
     #[error("I/O error")]
-    IoError(#[from] io::Error),
+    Io(#[from] io::Error),
     #[error("database error")]
-    DatabaseError(#[from] DatabaseError),
+    Database(#[from] DatabaseError),
     #[error("genesis block error")]
-    GenesisBlockError(#[from] GenesisError),
+    GenesisBlock(#[from] GenesisError),
     #[error("pool build error")]
-    PoolBuildError(#[from] deadpool::managed::BuildError),
+    PoolBuild(#[from] deadpool::managed::BuildError),
     #[error("SQLite migration error")]
-    SqliteMigrationError(#[from] rusqlite_migration::Error),
+    SqliteMigration(#[from] rusqlite_migration::Error),
 }
 
 #[derive(Debug, Error)]
@@ -131,33 +114,14 @@ pub enum GenesisError {
     // ERRORS WITH AUTOMATIC CONVERSIONS FROM NESTED ERROR TYPES
     // ---------------------------------------------------------------------------------------------
     #[error("database error")]
-    DatabaseError(#[from] DatabaseError),
+    Database(#[from] DatabaseError),
     // TODO: Check if needed.
     #[error("block error")]
-    BlockError,
-    #[error("merkle error")]
-    MerkleError(#[from] MerkleError),
+    Block,
+    #[error("failed to build genesis account tree")]
+    AccountTree(#[source] AccountTreeError),
     #[error("failed to deserialize genesis file")]
-    GenesisFileDeserializationError(#[from] DeserializationError),
-    #[error("retrieving genesis block header failed")]
-    SelectBlockHeaderByBlockNumError(#[from] Box<DatabaseError>),
-
-    // OTHER ERRORS
-    // ---------------------------------------------------------------------------------------------
-    #[error("apply block failed")]
-    ApplyBlockFailed(#[source] InteractError),
-    #[error("failed to read genesis file \"{genesis_filepath}\"")]
-    FailedToReadGenesisFile {
-        genesis_filepath: String,
-        source: io::Error,
-    },
-    #[error(
-        "block header in store doesn't match block header in genesis file. Expected {expected_genesis_header:?}, but store contained {block_header_in_store:?}"
-    )]
-    GenesisBlockHeaderMismatch {
-        expected_genesis_header: Box<BlockHeader>,
-        block_header_in_store: Box<BlockHeader>,
-    },
+    GenesisFileDeserialization(#[from] DeserializationError),
 }
 
 // ENDPOINT ERRORS
@@ -182,6 +146,10 @@ pub enum InvalidBlockError {
     NewBlockInvalidNullifierRoot,
     #[error("new block `prev_block_commitment` must match the chain's tip")]
     NewBlockInvalidPrevCommitment,
+    #[error("nullifier in new block is already spent")]
+    NewBlockNullifierAlreadySpent(#[source] NullifierTreeError),
+    #[error("duplicate account ID prefix in new block")]
+    NewBlockDuplicateAccountIdPrefix(#[source] AccountTreeError),
 }
 
 #[derive(Error, Debug)]