+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v0.9.0 (TBD)
+
+### Changes
+
+- [BREAKING] Simplified node bootstrapping (#776).
+  - Database is now created during bootstrap process instead of on first startup.
+  - Data directory is no longer created but is instead expected to exist.
+  - The genesis block can no longer be configured which also removes the `store dump-genesis` command.
+- [BREAKING] Use `AccountTree` and update account witness proto definitions (#783).
+- Enabled running RPC component in `read-only` mode (#802).
+- [BREAKING] Update name of `ChainMmr` to `PartialBlockchain` (#807).
+
 ## v0.8.0 (2025-03-26)
 
 ### Enhancements